refactor: 응답 메시지를 생성, 전송하는 역할 분리
- 핸들러의 동작 결과를 받아 HTTP 응답 메시지를 생성, 전송하는 클래스 추가
